+++ src/sox.h
@@ -9,7 +9,14 @@
 
 
 
-int8_t sox_init( float *cur, uint16_t *vol, int16_t *tmp, uint16_t *soc, uint16_t *grpSOC , float *soh, float *grpSOH, uint16_t *cycleCount);
+int8_t sox_init( 
+    float *cur, 
+    uint16_t *vol, 
+    int16_t *tmp, 
+    uint16_t *soc, 
+    uint16_t *grpSOC , 
+    double *soh, double *grpSOH, 
+    uint32_t *cycleCount);
 int8_t sox_task(bool full, bool empty);
 #endif
